% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_device_image}
\alias{get_device_image}
\title{Title}
\usage{
get_device_image(which = dev.cur(),
  image_height = pkg_options("image_height"),
  image_aspr = pkg_options("image_aspr"),
  image_res = pkg_options("image_res"))
}
\description{
Title
}
