% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pkg_options}
\alias{pkg_options}
\title{Set or get options for my package}
\usage{
pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{print_code}}{(\code{logical};TRUE) Print R code to the tweet? }
 \item{\code{print_output}}{(\code{logical};FALSE) Print R output to the tweet? }
 \item{\code{do_gist}}{(\code{logical};TRUE) Create a gist, and link to it? }
 \item{\code{tweet_image}}{(\code{logical};TRUE) Add an image to the tweet? }
 \item{\code{gif_from_images}}{(\code{logical};TRUE) If multiple plots, create an animated GIF? }
 \item{\code{gif_delay}}{(\code{numeric};2) Delay between GIF frame transitions (seconds) }
 \item{\code{image_height}}{(\code{numeric};800) Image height, in pixels }
 \item{\code{image_aspr}}{(\code{numeric};2) Image aspect ratio }
 \item{\code{image_res}}{(\code{numeric};200) Image resolution }
}
}

