% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{post_and_return_id}
\alias{post_and_return_id}
\title{Post a tweet and get its ID}
\usage{
post_and_return_id(...)
}
\arguments{
\item{...}{Arguments to pass to rtweet::post_tweet}
}
\description{
Post a tweet and get its ID
}
