% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_a_blog.R
\name{split_tweet0}
\alias{split_tweet0}
\title{Initial user-defined splits, then passes each segment off}
\usage{
split_tweet0(s, split_on = "\\\\n---*\\\\n", leave_space, max_char)
}
\arguments{
\item{s}{string containing tweet}

\item{split_on}{User-defined POSIX regular expression for forced splits in tweets}

\item{leave_space}{integer; how much space to leave at the end for numbering}

\item{max_char}{maximum characters that can go in a tweet}
}
\description{
Initial user-defined splits, then passes each segment off
}
