% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_code.R
\name{tweet_code}
\alias{tweet_code}
\title{Tweet R code}
\usage{
tweet_code(code, pre_text = NULL, reply = NULL,
  print_code = pkg_options("print_code"),
  print_output = pkg_options("print_output"),
  print_errors = pkg_options("print_errors"),
  stop_on_errors = pkg_options("stop_on_errors"),
  do_gist = pkg_options("do_gist"), do_tweet = TRUE,
  tweet_image = pkg_options("tweet_image"),
  gif_from_images = pkg_options("gif_from_images"),
  gif_delay = pkg_options("gif_delay"),
  image_height = pkg_options("image_height"),
  image_aspr = pkg_options("image_aspr"),
  image_res = pkg_options("image_res"), envir = parent.frame())
}
\arguments{
\item{code}{string variable containing R code}

\item{pre_text}{Text to prepend to the tweet (null if none)}

\item{print_code}{print the code in the tweet?}

\item{print_output}{print the output in the tweet?}

\item{do_gist}{create a GitHub gist with the code?}

\item{tweet_image}{Tweet an image with the code (if code creates image)?}

\item{gif_from_images}{If code creates multiple images, combine them as GIF?}

\item{gif_delay}{delay, in seconds, between GIF frames}

\item{image_height}{height of images in pixels}

\item{image_aspr}{aspect ratio of images}

\item{image_res}{resolution of images}

\item{envir}{environment in which to evaluate the R code}
}
\description{
Tweet R code
}
