% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_device.R
\name{tweet_device}
\alias{tweet_device}
\title{Title}
\usage{
tweet_device(which = dev.cur(), tweet_text = "", reply = NULL,
  do_tweet = TRUE, image_height = pkg_options("image_height"),
  image_aspr = pkg_options("image_aspr"),
  image_res = pkg_options("image_res"),
  image_res_scale = pkg_options("image_res_scale"),
  open_browser = pkg_options("open_browser"))
}
\description{
Title
}
