\name{tweetRcode-package}
\alias{tweetRcode-package}
\alias{tweetRcode}
\docType{package}
\title{
\packageTitle{tweetRcode}
}
\description{
\packageDescription{tweetRcode}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tweetRcode}
\packageIndices{tweetRcode}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tweetRcode}

Maintainer: \packageMaintainer{tweetRcode}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
