% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getablog.R
\name{get_a_blog}
\alias{get_a_blog}
\title{Tweet, potentially creating tweet storms by splitting tweets on white space}
\usage{
get_a_blog(s, leave_space = pkg_options("getablog_leave_space"),
  max_char = pkg_options("getablog_max_char"), do_tweet = TRUE)
}
\arguments{
\item{s}{character string; text of tweet(s)}

\item{leave_space}{integer; how much space to leave at the end for numbering}

\item{max_char}{maximum characters that can go in a tweet}

\item{do_tweet}{logical; tweet, or just return the splitted text?}
}
\description{
Tweet, potentially creating tweet storms by splitting tweets on white space
}
